Move upvoter-specific things out of background.ts
The long term goal is to define a clear interface that can be
implemented via both the Chrome extension API as well as Alfred via
cookiecutter-multicli.  I'm not there yet, but this will help share
the code in the meantime.
